requirements.txt
airtable-python-wrapper==0.15.2
python-dateutil 
beautifulsoup4 
pandas 
openpyxl 
xlsxwriter
loguru
aiogram
python-dotenv